+++ src/components/Search/index.js
@@ -1,6 +1,6 @@
 import React, { PureComponent } from 'react';
 import { connect } from 'react-redux'
-import { fetchCities } from '../actions'
+import { fetchCities } from '../../actions'
 import SearchResults from '../SearchResults'
 import './styles.css'
