+++ src/components/Search/styles.css
