+++ src/components/SearchHistory/styles.css
