+++ src/components/SearchResults/styles.css
